AzureBicepCosmosDB -> AzureCosmosDB.
