+++ playground/CosmosEndToEnd/CosmosEndToEnd.AppHost/Properties/launchSettings.json
@@ -11,18 +11,18 @@
         "DOTNET_ENVIRONMENT": "Development",
         "DOTNET_DASHBOARD_OTLP_ENDPOINT_URL": "http://localhost:16175",
         "DOTNET_ASPIRE_SHOW_DASHBOARD_RESOURCES": "true"
-      },
-      "generate-manifest": {
-        "commandName": "Project",
-        "launchBrowser": true,
-        "dotnetRunMessages": true,
-        "commandLineArgs": "--publisher manifest --output-path aspire-manifest.json",
-        "applicationUrl": "http://localhost:15888",
-        "environmentVariables": {
-          "ASPNETCORE_ENVIRONMENT": "Development",
-          "DOTNET_ENVIRONMENT": "Development",
-          "DOTNET_DASHBOARD_OTLP_ENDPOINT_URL": "http://localhost:16175"
-        }
+      }
+    },
+    "generate-manifest": {
+      "commandName": "Project",
+      "launchBrowser": true,
+      "dotnetRunMessages": true,
+      "commandLineArgs": "--publisher manifest --output-path aspire-manifest.json",
+      "applicationUrl": "http://localhost:15888",
+      "environmentVariables": {
+        "ASPNETCORE_ENVIRONMENT": "Development",
+        "DOTNET_ENVIRONMENT": "Development",
+        "DOTNET_DASHBOARD_OTLP_ENDPOINT_URL": "http://localhost:16175"
       }
     }
   }
