+++ playground/CosmosEndToEnd/CosmosEndToEnd.AppHost/aspire-manifest.json
@@ -1,0 +1,41 @@
+{
+  "resources": {
+    "cosmos": {
+      "type": "azure.bicep.v0",
+      "connectionString": "{cosmos.secretOutputs.connectionString}",
+      "path": "aspire.hosting.azure.bicep.cosmosdb.bicep",
+      "params": {
+        "databaseAccountName": "cosmos",
+        "databases": [
+          "db"
+        ],
+        "keyVaultName": ""
+      }
+    },
+    "db": {
+      "type": "value.v0",
+      "connectionString": "{cosmos.connectionString}"
+    },
+    "api": {
+      "type": "project.v0",
+      "path": "../CosmosEndToEnd.ApiService/CosmosEndToEnd.ApiService.csproj",
+      "env": {
+        "OTEL_DOTNET_EXPERIMENTAL_OTLP_EMIT_EXCEPTION_LOG_ATTRIBUTES": "true",
+        "OTEL_DOTNET_EXPERIMENTAL_OTLP_EMIT_EVENT_LOG_ATTRIBUTES": "true",
+        "ConnectionStrings__db": "{db.connectionString}"
+      },
+      "bindings": {
+        "http": {
+          "scheme": "http",
+          "protocol": "tcp",
+          "transport": "http"
+        },
+        "https": {
+          "scheme": "https",
+          "protocol": "tcp",
+          "transport": "http"
+        }
+      }
+    }
+  }
+}
