+++ src/Aspire.Hosting.Azure/AzureCosmosDBEmulatorResource.cs
@@ -2,7 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using Aspire.Hosting.ApplicationModel;
-using Aspire.Hosting.Azure.Data.Cosmos;
 
 namespace Aspire.Hosting.Azure;
