+++ src/Aspire.Hosting.Azure/AzureCosmosDBResource.cs
@@ -12,7 +12,7 @@ namespace Aspire.Hosting;
 /// <summary>
 /// A resource that represents an Azure Cosmos DB.
 /// </summary>
-public class AzureBicepCosmosDBResource(string name) :
+public class AzureCosmosDBResource(string name) :
     AzureBicepResource(name, templateResouceName: "Aspire.Hosting.Azure.Bicep.cosmosdb.bicep"),
     IResourceWithConnectionString
 {
@@ -58,15 +58,15 @@ public class AzureBicepCosmosDBResource(string name) :
 /// <summary>
 /// A resource that represents an Azure Cosmos DB database.
 /// </summary>
-public class AzureBicepCosmosDBDatabaseResource(string name, AzureBicepCosmosDBResource cosmosDB) :
+public class AzureCosmosDBDatabaseResource(string name, AzureCosmosDBResource cosmosDB) :
     Resource(name),
     IResourceWithConnectionString,
-    IResourceWithParent<AzureBicepCosmosDBResource>
+    IResourceWithParent<AzureCosmosDBResource>
 {
     /// <summary>
     /// Gets the parent Azure Cosmos DB resource.
     /// </summary>
-    public AzureBicepCosmosDBResource Parent => cosmosDB;
+    public AzureCosmosDBResource Parent => cosmosDB;
 
     /// <summary>
     /// Gets the connection string template for the manifest for the Azure Cosmos DB database resource.
@@ -83,16 +83,15 @@ public class AzureBicepCosmosDBDatabaseResource(string name, AzureBicepCosmosDBR
 
     internal void WriteToManifest(ManifestPublishingContext context)
     {
-        context.Writer.WriteString("type", "azure.bicep.v0");
+        context.Writer.WriteString("type", "value.v0");
         context.Writer.WriteString("connectionString", ConnectionStringExpression);
-        context.Writer.WriteString("parent", Parent.Name);
     }
 }
 
 /// <summary>
 /// Extension methods for adding Azure Cosmos DB resources to the application model.
 /// </summary>
-public static class AzureBicepCosmosExtensions
+public static class AzureCosmosExtensions
 {
     /// <summary>
     /// Adds an Azure Cosmos DB connection to the application model.
@@ -100,9 +99,9 @@ public static class AzureBicepCosmosExtensions
     /// <param name="builder">The <see cref="IDistributedApplicationBuilder"/>.</param>
     /// <param name="name">The name of the resource. This name will be used as the connection string name when referenced in a dependency.</param>
     /// <returns>A reference to the <see cref="IResourceBuilder{T}"/>.</returns>
-    public static IResourceBuilder<AzureBicepCosmosDBResource> AddBicepCosmosDb(this IDistributedApplicationBuilder builder, string name)
+    public static IResourceBuilder<AzureCosmosDBResource> AddAzureCosmosDB(this IDistributedApplicationBuilder builder, string name)
     {
-        var resource = new AzureBicepCosmosDBResource(name);
+        var resource = new AzureCosmosDBResource(name);
         return builder.AddResource(resource)
                       .WithParameter("databaseAccountName", resource.CreateBicepResourceName())
                       .WithParameter("databases", resource.Databases)
@@ -111,21 +110,43 @@ public static class AzureBicepCosmosExtensions
     }
 
     /// <summary>
-    /// Configures an Azure Cosmos DB resource to be emulated using the Azure Cosmos DB emulator with the NoSQL API. This resource requires an <see cref="AzureBicepCosmosDBResource"/> to be added to the application model.
+    /// Configures an Azure Cosmos DB resource to be emulated using the Azure Cosmos DB emulator with the NoSQL API. This resource requires an <see cref="AzureCosmosDBResource"/> to be added to the application model.
     /// For more information on the Azure Cosmos DB emulator, see <a href="https://learn.microsoft.com/azure/cosmos-db/emulator#authentication"></a>
     /// </summary>
     /// <param name="builder">The Azure Cosmos DB resource builder.</param>
-    /// <param name="port">The port used for the client SDK to access the emulator. Defaults to <c>8081</c></param>
-    /// <param name="imageTag">The image tag for the <c>mcr.microsoft.com/cosmosdb/linux/azure-cosmos-emulator</c> image.</param>
+    /// <param name="configureContainer">Callback that exposes underlying container used for emulation to allow for customization.</param>
     /// <returns>A reference to the <see cref="IResourceBuilder{T}"/>.</returns>
     /// <remarks>
     /// When using the Azure Cosmos DB emulator, the container requires a TLS/SSL certificate.
     /// For more information, see <a href="https://learn.microsoft.com/azure/cosmos-db/how-to-develop-emulator?tabs=docker-linux#export-the-emulators-tlsssl-certificate"></a>
     /// </remarks>
-    public static IResourceBuilder<AzureBicepCosmosDBResource> UseEmulator(this IResourceBuilder<AzureBicepCosmosDBResource> builder, int? port = null, string? imageTag = null)
+    public static IResourceBuilder<AzureCosmosDBResource> UseEmulator(this IResourceBuilder<AzureCosmosDBResource> builder, Action<IResourceBuilder<AzureCosmosDBEmulatorResource>>? configureContainer = null)
     {
-        return builder.WithAnnotation(new EndpointAnnotation(ProtocolType.Tcp, name: "emulator", port: port, containerPort: 8081))
-                      .WithAnnotation(new ContainerImageAnnotation { Image = "mcr.microsoft.com/cosmosdb/linux/azure-cosmos-emulator", Tag = imageTag ?? "latest" });
+        builder.WithAnnotation(new EndpointAnnotation(ProtocolType.Tcp, name: "emulator", containerPort: 8081))
+               .WithAnnotation(new ContainerImageAnnotation { Image = "mcr.microsoft.com/cosmosdb/linux/azure-cosmos-emulator", Tag = "latest" });
+
+        if (configureContainer != null)
+        {
+            var surrogate = new AzureCosmosDBEmulatorResource(builder.Resource);
+            var surrogateBuilder = builder.ApplicationBuilder.CreateResourceBuilder(surrogate);
+            configureContainer(surrogateBuilder);
+        }
+
+        return builder;
+    }
+
+    /// <summary>
+    /// Configures the gateway port for the Azure Cosmos DB emulator.
+    /// </summary>
+    /// <param name="builder">Builder for the Cosmos emulator container</param>
+    /// <param name="port">Host port to bind to the emulator gateway port.</param>
+    /// <returns>Cosmos emulator resource builder.</returns>
+    public static IResourceBuilder<AzureCosmosDBEmulatorResource> UseGatewayPort(this IResourceBuilder<AzureCosmosDBEmulatorResource> builder, int? port)
+    {
+        return builder.WithEndpoint("emulator", endpoint =>
+        {
+            endpoint.Port = port;
+        });
     }
 
     /// <summary>
@@ -135,11 +156,11 @@ public static class AzureBicepCosmosExtensions
     /// <param name="name">The name of the resource.</param>
     /// <param name="databaseName">Name of database.</param>
     /// <returns>A reference to the <see cref="IResourceBuilder{T}"/>.</returns>
-    public static IResourceBuilder<AzureBicepCosmosDBDatabaseResource> AddDatabase(this IResourceBuilder<AzureBicepCosmosDBResource> builder, string name, string? databaseName = null)
+    public static IResourceBuilder<AzureCosmosDBDatabaseResource> AddDatabase(this IResourceBuilder<AzureCosmosDBResource> builder, string name, string? databaseName = null)
     {
         var dbName = databaseName ?? name;
 
-        var resource = new AzureBicepCosmosDBDatabaseResource(name, builder.Resource);
+        var resource = new AzureCosmosDBDatabaseResource(name, builder.Resource);
 
         builder.Resource.Databases.Add(dbName);
